+++ R/acled.api.R
@@ -18,13 +18,13 @@
 #' below for information on how the arguments "country" and "region" interact.
 #' @param region numeric or character vector. Supply one or more region codes (numeric) or region names (character)
 #' to narrow down which events should be retrieved. You can run `get.api.regions()` to view supported region codes and names,
-#' or visit [ACLED's resource library](https://acleddata.com/resources/general-guides/).
+#' or see [ACLED's API Guide](https://apidocs.acleddata.com/).
 #' See the details below for information on how the arguments "country" and "region" interact.
 #' @param start.date character string. Supply the earliest date to be retrieved. Format: "yyyy-mm-dd".
 #' @param end.date character string. Supply the last date to be retrieved. Format: "yyyy-mm-dd".
 #' @param add.variables character vector. Supply the names of ACLED variables you wish to add to the
 #' default output (see [ACLED's codebook](https://acleddata.com/resources/general-guides/) for details). The default
-#' output includes: region, country, year, event_date, source, admin1, admin2, admin3, location, latitude, longitude, event_type, sub_event_type,
+#' output includes: event_id_cnty, region, country, year, event_date, source, admin1, admin2, admin3, location, latitude, longitude, event_type, sub_event_type,
 #' interaction, fatalities, tags, and the download timestamp.
 #' @param all.variables logical. When set to FALSE (default), a narrow default selection of variables is returned (which
 #' can be refined using the argument add.variables). If set to TRUE, all variables are included in the output (overrides
@@ -35,21 +35,29 @@
 #' retrieved (see [ACLED's codebook](https://acleddata.com/resources/general-guides/) for details.
 #' @param other.query character vector. Allows users to add their own ACLED API queries to the
 #' GET call. Vector elements are assumed to be individual queries, and are automatically separated by an & sign.
+#' See [ACLED's API Guide](https://apidocs.acleddata.com/) for information on making your custom query.
 #' @details The function _`acled.api()`_ is an R wrapper for
 #' the [Armed Conflict Location & Event Data Project](https://acleddata.com/) API.
 #' Internally it uses _`httr`_ to access the API, and _`jsonlite`_ to manage the JSON content that the call returns. The JSON data
 #' are converted into the base class _`data.frame`_. Variables are of class _`character`_ by default.
 #' Variables which only contain numbers as recognized by the regular
 #' expression `^[0-9]+$` are coerced into _`numeric`_ before the _`data.frame`_ object is returned. \cr \cr
-#' The user's registered email address and ACLED access key can be supplied as strings directly to their respective arguments,
+#' *Authentification:* The user's registered email address and ACLED access key, which they obtain after registering with [ACLED access](https://developer.acleddata.com/), can be supplied as strings directly to their respective arguments,
 #' or set in advance as environment variables
-#' using \cr \cr _`Sys.setenv(ACLED_EMAIL_ADDRESS="your.email.address")`_ and \cr \cr _`Sys.setenv(ACLED_ACCESS_KEY="your.access.key")`_. \cr \cr
-#' If both the country argument and the region argument are NULL (default), all available countries are retrieved. The same applies to
+#' using \cr _`Sys.setenv(ACLED_EMAIL_ADDRESS="your.email.address")`_ and \cr _`Sys.setenv(ACLED_ACCESS_KEY="your.access.key")`_. \cr \cr
+#' *Retrieving all data at once:* If both the country argument and the region argument are NULL (default), all available countries are retrieved. The same applies to
 #' the time frame when both the start date and the end date are NULL (default). Note that the API cannot handle requests with only one
-#' of the dates specified (either none of them or both of them need to be supplied). \cr \cr
-#' The ACLED API combines the country argument and the region argument with a logical AND operator. Therefore, specifying e.g. the
+#' of the dates specified (either none of them or both of them need to be supplied). More recent versions of the API implemented a bandwidth limit that prevents
+#' users from retrieving larger amounts of data in a single call. At the time of this writing (v. _`acled.api()`_ 1.1.8), ACLED recommends using pagination to retrieve data
+#' in separate calls of 5000 rows each. This is much less than what is actually possible to retrieve in a single call, and with a full dataset of >2 million rows, pagination is less
+#' practicable than simply downloading by individual regions or time periods. The separate data chunks obtained this way can then be combined using rbind(). Users who want to make use of
+#' pagination can do so using the _`other.query`_ filter. \cr \cr
+#' *Filter combinations:* By default, _`acled.api()`_ combines different filters with a logical AND operator.
+#' This usually conforms with desired behavior, with the country argument and the region argument as important exception: for example, specifying the
 #' country "Togo" and the region "Southern Africa" leads the API to query for a country named "Togo" in the region "Southern Africa".
-#' In this case, no data will be returned as no events match this query.
+#' In this case, no data will be returned as no events match this query, and it is recommended to make two separate calls, one for Togo and one for Southern Africa.
+#' More recent versions of the API also support OR operators to separate filters. These can be implemented through _`acled.api()`_ by using the _`other.query`_ filter.
+#' To do so, see the [ACLED API Guide](https://apidocs.acleddata.com/) for details.
 #' @return A data frame containing ACLED events.
 #' @import jsonlite
 #' @import httr
@@ -133,7 +141,7 @@ acled.api <- function(
   # country argument
   if ( (!is.null(country) & !is.character(country)) == TRUE ) {
     stop('If you wish to specify country names, these need to be supplied as a character vector.
-    Usage example: \n
+Usage example:
          acled.api(country = c("Kenya", "Togo"), start.date = "2004-08-20", end.date = "2005-05-15")', call. = FALSE)
   }
   if(is.character(country) == TRUE){
@@ -146,9 +154,9 @@ acled.api <- function(
   # region argument
   if ( (!is.null(region) & !is.numeric(region) & !is.character(region)) == TRUE ) {
     stop('If you wish to specify regions, these need to be supplied as a numeric vector (region codes) or character
-    vector (region names). See the ACLED API guide for exact region names and codes.
-    Usage example: \n
-         acled.api(region = c(1,2), start.date = "2004-08-20", end.date = "2005-05-15") or \n
+vector (region names). See the ACLED API guide for exact region names and codes.
+Usage example:
+         acled.api(region = c(1,2), start.date = "2004-08-20", end.date = "2005-05-15") or
          acled.api(region = c("Western Africa", "Middle Africa"), start.date = "2004-08-20", end.date = "2005-05-15")', call. = FALSE)
   }
   region.data.frame <- get.api.regions()[[1]]
@@ -162,7 +170,7 @@ acled.api <- function(
                      " supplied in argument 'region' ",
                      ifelse(length(invalid.region) > 1, "do", "does"),
                      " not match the original ACLED region codes.\n",
-                     "Check your spelling, and see the ACLED resource library",
+                     "Check your spelling, and see the ACLED API guide",
                      " for the correct codes or run get.api.regions()."), call. = FALSE)
       }
     region1 <- paste0("&region=", paste(region, collapse = "|") )
@@ -178,7 +186,7 @@ acled.api <- function(
                          " supplied in argument 'region' ",
                          ifelse(length(invalid.region) > 1, "do", "does"),
                          " not match the original ACLED region names.\n",
-                         "Check your spelling, and see the ACLED resource library",
+                         "Check your spelling, and see the ACLED API guide",
                          " for the correct names or run get.api.regions()."), call. = FALSE)
          }
   }
@@ -189,9 +197,8 @@ acled.api <- function(
   # start.date & end.date arguments
   if ( (is.null(start.date) & is.null(end.date)) == FALSE ){
     if( (is.null(start.date) | is.null(end.date)) == TRUE){
-    stop("You need to supply either no start date and no end date, in which case all available dates are requested, or both
-    a start date and an end date.
-    Usage example: \n
+    stop("You need to supply either no start date and no end date, in which case all available dates are requested, or both a start date and an end date.
+Usage example:
          acled.api(region = c(1), start.date = '1995-01-15', end.date = '2005-12-15') or
          acled.api(region = c(1), start.date = NULL, end.date = NULL)", call. = FALSE)
     }
@@ -201,20 +208,13 @@ acled.api <- function(
   }
   time.frame1 <- paste0("&event_date=", paste(start.date, end.date, sep = "|"), "&event_date_where=BETWEEN")
 
-  if ( (is.null(start.date) & is.null(end.date) & is.null(region) & is.null(country)) == TRUE ) {
-    warning('You are trying to download data for all available regions and time periods.
-    ACLED has implemented bandwidth restrictions that may prevent you from doing so.
-    Therefore, if you are encountering an error, try to make separate calls by regions or time periods.
-    After downloading the data in these chunks, you can then combine them with rbind().', call. = FALSE)
-  }
-
   # add.variables and all.variables argument
   if( is.logical(all.variables)==TRUE ){
     if( all.variables==FALSE ){
         if( is.null(add.variables)==TRUE ){
-          variables <- "&fields=region|country|year|event_date|source|admin1|admin2|admin3|location|latitude|longitude|event_type|sub_event_type|interaction|fatalities|tags|timestamp"
+          variables <- "&fields=event_id_cnty|region|country|year|event_date|source|admin1|admin2|admin3|location|latitude|longitude|event_type|sub_event_type|interaction|fatalities|tags|timestamp"
           }else{
-            variables <- paste0("&fields=region|country|year|event_date|source|admin1|admin2|admin3|location|latitude|longitude|event_type|sub_event_type|interaction|fatalities|tags|timestamp",
+            variables <- paste0("&fields=event_id_cnty|region|country|year|event_date|source|admin1|admin2|admin3|location|latitude|longitude|event_type|sub_event_type|interaction|fatalities|tags|timestamp",
                         "|", paste(add.variables, collapse = "|") )
             }
       }else{
@@ -255,13 +255,20 @@ acled.api <- function(
   # other.query argument
   other.query1 <- ifelse( is.null(other.query)==TRUE, "", paste0("&", paste(other.query, collapse = "&")) )
 
+  # bandwidth limit warning
+  if ( (is.null(start.date) & is.null(end.date) & is.null(region) & is.null(country)) == TRUE ) {
+    warning('You are trying to download data for all available regions and time periods.
+ACLED has bandwidth restrictions that may prevent you from doing so.
+This can be solved by making several smaller calls, separately downloading individual regions or time periods.
+After downloading the data in these chunks, you can then combine them with rbind().', call. = FALSE)
+  }
 
   # ACLED ping
   acled.url.ping <- tryCatch(
     httr::GET("https://api.acleddata.com/"),
     error = function(e) e )
   if( any(class(acled.url.ping) == "error") ) {
-    message("The resource api.acleddata.com cannot be reached. \n
+    message("The resource api.acleddata.com cannot be reached.
     1. Please check your internet connection.
     2. If the internet connection is reliable, the server may be temporarily unavailable; in this case please try again later.
     3. If the problem persists, please contact the package maintainer as the resource may have moved.")
@@ -276,13 +283,15 @@ acled.api <- function(
   response <- httr::GET(url)
   if ( exists("response")==FALSE ) {
     message("GET request was unsuccessful. Please check your internet connection. If the problem persists despite a reliable internet connection,
-    the server may be temporarily not available; in this case try again later.",
-    call. = FALSE)
+the server may be temporarily not available; in this case try again later.")
     return(NULL)
   }
   if (httr::http_type(response) != "application/json" | length(response$content)==0) {
     message(paste0("GET request was unsuccessful, giving the status code ",
-                response$status_code, ". \n For information on HTTP status codes, see https://developer.mozilla.org/en-US/docs/Web/HTTP/Status"), call. = FALSE)
+                response$status_code, ". \n For information on status codes, see https://apidocs.acleddata.com/ and https://developer.mozilla.org/en-US/docs/Web/HTTP/Status.
+                \n If you are trying to download a large dataset (e.g., multiple regions over time), you may be hitting ACLED bandwidth restrictions.
+  This can be solved by making several smaller calls, separately downloading individual regions or time periods.
+  After downloading the data in these chunks, you can then combine them with rbind()."))
     return(NULL)
   }
 
@@ -296,8 +305,8 @@ acled.api <- function(
   json.content <- json.content$data
 
   if( length(json.content)==0L ){
-    message("No data found for this area and time period.
-            Or did you supply both countries and (other) regions? These cannot be combined.")
+    message("No data found for this area, time period, and/or page.
+Or did you supply both countries and (other) regions? These cannot be combined.")
     return(NULL)
   }